+++ src/components/mypage/MyBookmarkIcon.tsx
@@ -2,8 +2,8 @@ function MyBookmarkIcon() {
   return (
     <svg
       xmlns="http://www.w3.org/2000/svg"
-      width="25"
-      height="25"
+      width="26"
+      height="26"
       viewBox="0 0 24 24"
       fill="none"
     >
